+++ src/web/assets/cp/CpAsset.php
@@ -12,6 +12,7 @@ use craft\base\ElementInterface;
 use craft\config\GeneralConfig;
 use craft\elements\User;
 use craft\helpers\Assets;
+use craft\helpers\Cp;
 use craft\helpers\Json;
 use craft\helpers\StringHelper;
 use craft\helpers\UrlHelper;
@@ -383,7 +384,7 @@ JS;
             'right' => $orientation === 'ltr' ? 'right' : 'left',
             'runQueueAutomatically' => $generalConfig->runQueueAutomatically,
             'scriptName' => basename($request->getScriptFile()),
-            'siteId' => $upToDate ? (int)$sitesService->currentSite->id : null,
+            'siteId' => $upToDate ? (Cp::requestedSite()->id ?? null) : null,
             'sites' => $this->_sites($sitesService),
             'siteToken' => $generalConfig->siteToken,
             'slugWordSeparator' => $generalConfig->slugWordSeparator,